move Schema metadata to correct namespace
