+++ src/Sulu/Bundle/AdminBundle/Tests/Unit/Metadata/Schema/SchemaTest.php
@@ -9,11 +9,11 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Sulu\Bundle\AdminBundle\Tests\Unit\ResourceMetadata\Schema;
+namespace Sulu\Bundle\AdminBundle\Tests\Unit\Metadata\Schema;
 
 use PHPUnit\Framework\TestCase;
-use Sulu\Bundle\AdminBundle\ResourceMetadata\Schema\Property;
-use Sulu\Bundle\AdminBundle\ResourceMetadata\Schema\Schema;
+use Sulu\Bundle\AdminBundle\Metadata\Schema\Property;
+use Sulu\Bundle\AdminBundle\Metadata\Schema\Schema;
 
 class SchemaTest extends TestCase
 {
